backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# AI Providers
openai>=1.30.0
google-genai>=1.0.0
anthropic>=0.25.0

# PDF parsing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.3.0
llama-parse>=0.5.0
liteparse>=1.2.0

# MongoDB
motor>=3.4.0
pymongo>=4.7.0

# Celery + Redis
celery>=5.3.6
redis>=5.0.4

# Config
pydantic>=2.7.0
pydantic-settings>=2.2.0
